+++ backend/package.json
@@ -13,6 +13,7 @@
     "express": "^4.18.2",
     "mongoose": "^8.2.1",
     "morgan": "~1.9.1",
-    "nodemon": "^3.0.3"
+    "nodemon": "^3.0.3",
+    "password-hash": "^1.2.2"
   }
 }
